Custom Effect and target creature fix
